requirements.txt
Flask==3.0.0
Pillow==10.1.0
Werkzeug==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
pydub==0.25.1
python-magic==0.4.27
requests==2.31.0
python-slugify==8.0.1
blinker==1.7.0
